Created Keyboard helper, removed Keyboard binding library from Keyboard extension
